+++ service/Controllers/productController.go
@@ -1,0 +1,75 @@
+package Controllers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"gorm.io/gorm"
+
+	"service/Config"
+	"service/Models"
+)
+
+func CreateProduct(c *gin.Context) {
+	var product Models.Product
+	c.BindJSON(&product)
+	var err = Models.CreateProduct(&product)
+	if err != nil {
+		c.AbortWithStatus(http.StatusInternalServerError)
+	} else {
+		c.JSON(http.StatusOK, product)
+	}
+}
+
+func GetProduct(c *gin.Context) {
+	var products []Models.Product
+	err := Models.GetProduct(&products)
+	if err != nil {
+		c.AbortWithStatus(http.StatusInternalServerError)
+	} else {
+		c.JSON(http.StatusOK, products)
+	}
+}
+
+func GetProductByID(c *gin.Context) {
+	id := c.Params.ByName("id")
+	var product Models.Product
+	err := Models.GetProductByID(&product, id)
+	if err != nil {
+		if err == gorm.ErrRecordNotFound {
+			c.AbortWithStatus(http.StatusNotFound)
+		} else {
+			c.AbortWithStatus(http.StatusInternalServerError)
+		}
+	} else {
+		c.JSON(http.StatusOK, product)
+	}
+}
+
+func UpdateProduct(c *gin.Context) {
+	var product Models.Product
+	id := c.Params.ByName("id")
+	err := Models.UpdateProduct(&product, id)
+	if err != nil {
+		c.JSON(http.StatusNotFound, product)
+	}
+	c.BindJSON(&product)
+	err = Config.DB.Save(&product).Error
+	if err != nil {
+		c.AbortWithStatus(http.StatusInternalServerError)
+	} else {
+		c.JSON(http.StatusOK, product)
+	}
+
+}
+
+func DeleteProduct(c *gin.Context) {
+	id := c.Param("id")
+	var product Models.Product
+	err := Models.DeleteProduct(&product, id)
+	if err != nil {
+		c.AbortWithStatus(http.StatusInternalServerError)
+	} else {
+		c.JSON(http.StatusOK, gin.H{"id": id, "message": "product deleted"})
+	}
+}
